refactor: convert files to ts
install typescript
