+++ package.json
@@ -17,10 +17,12 @@
     "dotenv": "^16.4.1",
     "express": "^4.18.2",
     "make-runnable": "^1.4.1",
-    "pg": "^8.11.3"
+    "pg": "^8.11.3",
+    "typescript": "^5.3.3"
   },
   "devDependencies": {
     "@types/body-parser": "^1.19.5",
+    "@types/express": "^4.17.21",
     "@types/pg": "^8.11.0",
     "nodemon": "^3.0.3"
   },
